requirements.txt
fastapi
uvicorn
httpx
python-dotenv
supabase
asyncio
transformers
torch
